Consolidate user login and authentication code
