+++ common/djangoapps/student/helpers.py
@@ -17,10 +17,6 @@ from django.contrib.auth.models import User
 from django.db import IntegrityError, transaction
 from django.utils import http
 from django.utils.translation import ugettext as _
-from oauth2_provider.models import AccessToken as dot_access_token
-from oauth2_provider.models import RefreshToken as dot_refresh_token
-from provider.oauth2.models import AccessToken as dop_access_token
-from provider.oauth2.models import RefreshToken as dop_refresh_token
 from pytz import UTC
 from six import iteritems, text_type
 import third_party_auth
@@ -370,16 +366,6 @@ def get_redirect_to(request):
                     break
 
     return redirect_to
-
-
-def destroy_oauth_tokens(user):
-    """
-    Destroys ALL OAuth access and refresh tokens for the given user.
-    """
-    dop_access_token.objects.filter(user=user.id).delete()
-    dop_refresh_token.objects.filter(user=user.id).delete()
-    dot_access_token.objects.filter(user=user.id).delete()
-    dot_refresh_token.objects.filter(user=user.id).delete()
 
 
 def generate_activation_email_context(user, registration):
