+++ common/djangoapps/student/views/__init__.py
@@ -4,5 +4,4 @@ Combines all of the broken out student views
 
 # pylint: disable=wildcard-import
 from dashboard import *
-from login import *
 from management import *
