+++ common/djangoapps/student/views/management.py
@@ -3,19 +3,14 @@ Student Views
 """
 
 import datetime
-import json
 import logging
 import uuid
-import warnings
 from collections import namedtuple
 
-import analytics
-import dogstats_wrapper as dog_stats_api
 from bulk_email.models import Optout
 from courseware.courses import get_courses, sort_by_announcement, sort_by_start_date
 from django.conf import settings
 from django.contrib import messages
-from django.contrib.auth import login as django_login
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import AnonymousUser, User
 from django.contrib.auth.views import password_reset_confirm
@@ -31,60 +26,52 @@ from django.template.context_processors import csrf
 from django.template.response import TemplateResponse
 from django.utils.encoding import force_bytes, force_text
 from django.utils.http import base36_to_int, urlsafe_base64_encode
-from django.utils.translation import get_language, ungettext
 from django.utils.translation import ugettext as _
 from django.views.decorators.csrf import csrf_exempt, ensure_csrf_cookie
-from django.views.decorators.http import require_GET, require_POST
+from django.views.decorators.http import require_GET, require_POST, require_http_methods
 from edx_django_utils import monitoring as monitoring_utils
 from eventtracking import tracker
 from ipware.ip import get_ip
 # Note that this lives in LMS, so this dependency should be refactored.
-from notification_prefs.views import enable_notifications
 from opaque_keys import InvalidKeyError
 from opaque_keys.edx.keys import CourseKey
 from pytz import UTC
-from requests import HTTPError
-from six import text_type, iteritems
-from social_core.exceptions import AuthAlreadyAssociated, AuthException
-from social_django import utils as social_utils
+from six import text_type
 from xmodule.modulestore.django import modulestore
 
-import openedx.core.djangoapps.external_auth.views
-import third_party_auth
 import track.views
 from course_modes.models import CourseMode
+from edx_ace import ace
+from edx_ace.recipient import Recipient
 from edxmako.shortcuts import render_to_response, render_to_string
 from entitlements.models import CourseEntitlement
-from openedx.core.djangoapps.catalog.utils import (
-    get_programs_with_type,
-)
+
+from openedx.core.djangoapps.ace_common.template_context import get_base_template_context
+from openedx.core.djangoapps.catalog.utils import get_programs_with_type
 from openedx.core.djangoapps.embargo import api as embargo_api
-from openedx.core.djangoapps.external_auth.login_and_register import register as external_auth_register
-from openedx.core.djangoapps.lang_pref import LANGUAGE_KEY
+from openedx.core.djangoapps.oauth_dispatch.api import destroy_oauth_tokens
 from openedx.core.djangoapps.programs.models import ProgramsApiConfig
 from openedx.core.djangoapps.site_configuration import helpers as configuration_helpers
 from openedx.core.djangoapps.theming import helpers as theming_helpers
-from openedx.core.djangoapps.user_api import accounts as accounts_settings
-from openedx.core.djangoapps.user_api.accounts.utils import generate_password
+from openedx.core.djangoapps.theming.helpers import get_current_site
+from openedx.core.djangoapps.user_api.config.waffle import PREVENT_AUTH_USER_WRITES, SYSTEM_MAINTENANCE_MSG, waffle
+from openedx.core.djangoapps.user_api.errors import UserNotFound, UserAPIInternalError
 from openedx.core.djangoapps.user_api.models import UserRetirementRequest
 from openedx.core.djangoapps.user_api.preferences import api as preferences_api
-from openedx.core.djangoapps.user_api.config.waffle import PREVENT_AUTH_USER_WRITES, SYSTEM_MAINTENANCE_MSG, waffle
+
 from openedx.core.djangolib.markup import HTML, Text
 from openedx.features.journals.api import get_journals_context
-from student.cookies import set_logged_in_cookies
 from student.forms import AccountCreationForm, PasswordResetFormNoActive, get_registration_extension_form
 from student.helpers import (
     DISABLE_UNENROLL_CERT_STATES,
-    AccountValidationError,
     auth_pipeline_urls,
-    authenticate_new_user,
     cert_info,
     create_or_set_user_attribute_created_on_site,
-    destroy_oauth_tokens,
     do_create_account,
     generate_activation_email_context,
     get_next_url_for_login_page
 )
+from student.message_types import PasswordReset
 from student.models import (
     CourseEnrollment,
     PasswordHistory,
@@ -101,8 +88,6 @@ from student.models import (
 from student.signals import REFUND_ORDER
 from student.tasks import send_activation_email
 from student.text_me_the_app import TextMeTheAppFragmentView
-from third_party_auth import pipeline, provider
-from third_party_auth.saml import SAP_SUCCESSFACTORS_SAML_KEY
 from util.bad_request_rate_limiter import BadRequestRateLimiter
 from util.db import outer_atomic
 from util.json_request import JsonResponse
@@ -202,56 +187,6 @@ def index(request, extra_context=None, user=AnonymousUser()):
     return render_to_response('index.html', context)
 
 
-@ensure_csrf_cookie
-def register_user(request, extra_context=None):
-    """
-    Deprecated. To be replaced by :class:`student_account.views.login_and_registration_form`.
-    """
-    # Determine the URL to redirect to following login:
-    redirect_to = get_next_url_for_login_page(request)
-    if request.user.is_authenticated:
-        return redirect(redirect_to)
-
-    external_auth_response = external_auth_register(request)
-    if external_auth_response is not None:
-        return external_auth_response
-
-    context = {
-        'login_redirect_url': redirect_to,  # This gets added to the query string of the "Sign In" button in the header
-        'email': '',
-        'name': '',
-        'running_pipeline': None,
-        'pipeline_urls': auth_pipeline_urls(pipeline.AUTH_ENTRY_REGISTER, redirect_url=redirect_to),
-        'platform_name': configuration_helpers.get_value(
-            'platform_name',
-            settings.PLATFORM_NAME
-        ),
-        'selected_provider': '',
-        'username': '',
-    }
-
-    if extra_context is not None:
-        context.update(extra_context)
-
-    if context.get("extauth_domain", '').startswith(
-            openedx.core.djangoapps.external_auth.views.SHIBBOLETH_DOMAIN_PREFIX
-    ):
-        return render_to_response('register-shib.html', context)
-
-    # If third-party auth is enabled, prepopulate the form with data from the
-    # selected provider.
-    if third_party_auth.is_enabled() and pipeline.running(request):
-        running_pipeline = pipeline.get(request)
-        current_provider = provider.Registry.get_from_pipeline(running_pipeline)
-        if current_provider is not None:
-            overrides = current_provider.get_register_form_data(running_pipeline.get('kwargs'))
-            overrides['running_pipeline'] = running_pipeline
-            overrides['selected_provider'] = current_provider.name
-            context.update(overrides)
-
-    return render_to_response('register.html', context)
-
-
 def compose_and_send_activation_email(user, profile, user_registration=None):
     """
     Construct all the required params and send the activation email
@@ -277,6 +212,51 @@ def compose_and_send_activation_email(user, profile, user_registration=None):
         message_for_activation = ("Activation for %s (%s): %s\n" % (user, user.email, profile.name) +
                                   '-' * 80 + '\n\n' + message_for_activation)
     send_activation_email.delay(subject, message_for_activation, from_address, dest_addr)
+
+
+def send_reactivation_email_for_user(user):
+    try:
+        registration = Registration.objects.get(user=user)
+    except Registration.DoesNotExist:
+        return JsonResponse({
+            "success": False,
+            "error": _('No inactive user with this e-mail exists'),
+        })
+
+    try:
+        context = generate_activation_email_context(user, registration)
+    except ObjectDoesNotExist:
+        log.error(
+            u'Unable to send reactivation email due to unavailable profile for the user "%s"',
+            user.username,
+            exc_info=True
+        )
+        return JsonResponse({
+            "success": False,
+            "error": _('Unable to send reactivation email')
+        })
+
+    subject = render_to_string('emails/activation_email_subject.txt', context)
+    subject = ''.join(subject.splitlines())
+    message = render_to_string('emails/activation_email.txt', context)
+    from_address = configuration_helpers.get_value('email_from_address', settings.DEFAULT_FROM_EMAIL)
+    from_address = configuration_helpers.get_value('ACTIVATION_EMAIL_FROM_ADDRESS', from_address)
+
+    try:
+        user.email_user(subject, message, from_address)
+    except Exception:  # pylint: disable=broad-except
+        log.error(
+            u'Unable to send reactivation email from "%s" to "%s"',
+            from_address,
+            user.email,
+            exc_info=True
+        )
+        return JsonResponse({
+            "success": False,
+            "error": _('Unable to send reactivation email')
+        })
+
+    return JsonResponse({"success": True})
 
 
 @login_required
@@ -562,421 +542,6 @@ def user_signup_handler(sender, **kwargs):  # pylint: disable=unused-argument
             log.info(u'user {} originated from a white labeled "Microsite"'.format(kwargs['instance'].id))
 
 
-@transaction.non_atomic_requests
-def create_account_with_params(request, params):
-    """
-    Given a request and a dict of parameters (which may or may not have come
-    from the request), create an account for the requesting user, including
-    creating a comments service user object and sending an activation email.
-    This also takes external/third-party auth into account, updates that as
-    necessary, and authenticates the user for the request's session.
-
-    Does not return anything.
-
-    Raises AccountValidationError if an account with the username or email
-    specified by params already exists, or ValidationError if any of the given
-    parameters is invalid for any other reason.
-
-    Issues with this code:
-    * It is non-transactional except where explicitly wrapped in atomic to
-      alleviate deadlocks and improve performance. This means failures at
-      different places in registration can leave users in inconsistent
-      states.
-    * Third-party auth passwords are not verified. There is a comment that
-      they are unused, but it would be helpful to have a sanity check that
-      they are sane.
-    * The user-facing text is rather unfriendly (e.g. "Username must be a
-      minimum of two characters long" rather than "Please use a username of
-      at least two characters").
-    * Duplicate email raises a ValidationError (rather than the expected
-      AccountValidationError). Duplicate username returns an inconsistent
-      user message (i.e. "An account with the Public Username '{username}'
-      already exists." rather than "It looks like {username} belongs to an
-      existing account. Try again with a different username.") The two checks
-      occur at different places in the code; as a result, registering with
-      both a duplicate username and email raises only a ValidationError for
-      email only.
-    """
-    # Copy params so we can modify it; we can't just do dict(params) because if
-    # params is request.POST, that results in a dict containing lists of values
-    params = dict(params.items())
-
-    # allow to define custom set of required/optional/hidden fields via configuration
-    extra_fields = configuration_helpers.get_value(
-        'REGISTRATION_EXTRA_FIELDS',
-        getattr(settings, 'REGISTRATION_EXTRA_FIELDS', {})
-    )
-    # registration via third party (Google, Facebook) using mobile application
-    # doesn't use social auth pipeline (no redirect uri(s) etc involved).
-    # In this case all related info (required for account linking)
-    # is sent in params.
-    # `third_party_auth_credentials_in_api` essentially means 'request
-    # is made from mobile application'
-    third_party_auth_credentials_in_api = 'provider' in params
-
-    is_third_party_auth_enabled = third_party_auth.is_enabled()
-
-    if is_third_party_auth_enabled and (pipeline.running(request) or third_party_auth_credentials_in_api):
-        params["password"] = generate_password()
-
-    # in case user is registering via third party (Google, Facebook) and pipeline has expired, show appropriate
-    # error message
-    if is_third_party_auth_enabled and ('social_auth_provider' in params and not pipeline.running(request)):
-        raise ValidationError(
-            {'session_expired': [
-                _(u"Registration using {provider} has timed out.").format(
-                    provider=params.get('social_auth_provider'))
-            ]}
-        )
-
-    # if doing signup for an external authorization, then get email, password, name from the eamap
-    # don't use the ones from the form, since the user could have hacked those
-    # unless originally we didn't get a valid email or name from the external auth
-    # TODO: We do not check whether these values meet all necessary criteria, such as email length
-    do_external_auth = 'ExternalAuthMap' in request.session
-    if do_external_auth:
-        eamap = request.session['ExternalAuthMap']
-        try:
-            validate_email(eamap.external_email)
-            params["email"] = eamap.external_email
-        except ValidationError:
-            pass
-        if len(eamap.external_name.strip()) >= accounts_settings.NAME_MIN_LENGTH:
-            params["name"] = eamap.external_name
-        params["password"] = eamap.internal_password
-        log.debug(u'In create_account with external_auth: user = %s, email=%s', params["name"], params["email"])
-
-    extended_profile_fields = configuration_helpers.get_value('extended_profile_fields', [])
-    enforce_password_policy = not do_external_auth
-    # Can't have terms of service for certain SHIB users, like at Stanford
-    registration_fields = getattr(settings, 'REGISTRATION_EXTRA_FIELDS', {})
-    tos_required = (
-        registration_fields.get('terms_of_service') != 'hidden' or
-        registration_fields.get('honor_code') != 'hidden'
-    ) and (
-        not settings.FEATURES.get("AUTH_USE_SHIB") or
-        not settings.FEATURES.get("SHIB_DISABLE_TOS") or
-        not do_external_auth or
-        not eamap.external_domain.startswith(openedx.core.djangoapps.external_auth.views.SHIBBOLETH_DOMAIN_PREFIX)
-    )
-
-    form = AccountCreationForm(
-        data=params,
-        extra_fields=extra_fields,
-        extended_profile_fields=extended_profile_fields,
-        enforce_password_policy=enforce_password_policy,
-        tos_required=tos_required,
-    )
-    custom_form = get_registration_extension_form(data=params)
-
-    third_party_provider = None
-    running_pipeline = None
-    new_user = None
-
-    # Perform operations within a transaction that are critical to account creation
-    with outer_atomic(read_committed=True):
-        # first, create the account
-        (user, profile, registration) = do_create_account(form, custom_form)
-
-        # If a 3rd party auth provider and credentials were provided in the API, link the account with social auth
-        # (If the user is using the normal register page, the social auth pipeline does the linking, not this code)
-
-        # Note: this is orthogonal to the 3rd party authentication pipeline that occurs
-        # when the account is created via the browser and redirect URLs.
-
-        if is_third_party_auth_enabled and third_party_auth_credentials_in_api:
-            backend_name = params['provider']
-            request.social_strategy = social_utils.load_strategy(request)
-            redirect_uri = reverse('social:complete', args=(backend_name, ))
-            request.backend = social_utils.load_backend(request.social_strategy, backend_name, redirect_uri)
-            social_access_token = params.get('access_token')
-            if not social_access_token:
-                raise ValidationError({
-                    'access_token': [
-                        _("An access_token is required when passing value ({}) for provider.").format(
-                            params['provider']
-                        )
-                    ]
-                })
-            request.session[pipeline.AUTH_ENTRY_KEY] = pipeline.AUTH_ENTRY_REGISTER_API
-            pipeline_user = None
-            error_message = ""
-            try:
-                pipeline_user = request.backend.do_auth(social_access_token, user=user)
-            except AuthAlreadyAssociated:
-                error_message = _("The provided access_token is already associated with another user.")
-            except (HTTPError, AuthException):
-                error_message = _("The provided access_token is not valid.")
-            if not pipeline_user or not isinstance(pipeline_user, User):
-                # Ensure user does not re-enter the pipeline
-                request.social_strategy.clean_partial_pipeline(social_access_token)
-                raise ValidationError({'access_token': [error_message]})
-
-        # If the user is registering via 3rd party auth, track which provider they use
-        if is_third_party_auth_enabled and pipeline.running(request):
-            running_pipeline = pipeline.get(request)
-            third_party_provider = provider.Registry.get_from_pipeline(running_pipeline)
-
-        new_user = authenticate_new_user(request, user.username, params['password'])
-        django_login(request, new_user)
-        request.session.set_expiry(0)
-
-        if do_external_auth:
-            eamap.user = new_user
-            eamap.dtsignup = datetime.datetime.now(UTC)
-            eamap.save()
-            AUDIT_LOG.info(u"User registered with external_auth %s", new_user.username)
-            AUDIT_LOG.info(u'Updated ExternalAuthMap for %s to be %s', new_user.username, eamap)
-
-            if settings.FEATURES.get('BYPASS_ACTIVATION_EMAIL_FOR_EXTAUTH'):
-                log.info('bypassing activation email')
-                new_user.is_active = True
-                new_user.save()
-                AUDIT_LOG.info(
-                    u"Login activated on extauth account - {0} ({1})".format(new_user.username, new_user.email))
-
-    # Check if system is configured to skip activation email for the current user.
-    skip_email = skip_activation_email(
-        user, do_external_auth, running_pipeline, third_party_provider,
-    )
-
-    if skip_email:
-        registration.activate()
-    else:
-        compose_and_send_activation_email(user, profile, registration)
-
-    # Perform operations that are non-critical parts of account creation
-    create_or_set_user_attribute_created_on_site(user, request.site)
-
-    preferences_api.set_user_preference(user, LANGUAGE_KEY, get_language())
-
-    if settings.FEATURES.get('ENABLE_DISCUSSION_EMAIL_DIGEST'):
-        try:
-            enable_notifications(user)
-        except Exception:  # pylint: disable=broad-except
-            log.exception("Enable discussion notifications failed for user {id}.".format(id=user.id))
-
-    dog_stats_api.increment("common.student.account_created")
-
-    # Track the user's registration
-    if hasattr(settings, 'LMS_SEGMENT_KEY') and settings.LMS_SEGMENT_KEY:
-        tracking_context = tracker.get_tracker().resolve_context()
-        identity_args = [
-            user.id,
-            {
-                'email': user.email,
-                'username': user.username,
-                'name': profile.name,
-                # Mailchimp requires the age & yearOfBirth to be integers, we send a sane integer default if falsey.
-                'age': profile.age or -1,
-                'yearOfBirth': profile.year_of_birth or datetime.datetime.now(UTC).year,
-                'education': profile.level_of_education_display,
-                'address': profile.mailing_address,
-                'gender': profile.gender_display,
-                'country': text_type(profile.country),
-            }
-        ]
-
-        if hasattr(settings, 'MAILCHIMP_NEW_USER_LIST_ID'):
-            identity_args.append({
-                "MailChimp": {
-                    "listId": settings.MAILCHIMP_NEW_USER_LIST_ID
-                }
-            })
-
-        analytics.identify(*identity_args)
-
-        analytics.track(
-            user.id,
-            "edx.bi.user.account.registered",
-            {
-                'category': 'conversion',
-                'label': params.get('course_id'),
-                'provider': third_party_provider.name if third_party_provider else None
-            },
-            context={
-                'ip': tracking_context.get('ip'),
-                'Google Analytics': {
-                    'clientId': tracking_context.get('client_id')
-                }
-            }
-        )
-
-    # Announce registration
-    REGISTER_USER.send(sender=None, user=user, registration=registration)
-
-    create_comments_service_user(user)
-
-    try:
-        record_registration_attributions(request, new_user)
-    # Don't prevent a user from registering due to attribution errors.
-    except Exception:   # pylint: disable=broad-except
-        log.exception('Error while attributing cookies to user registration.')
-
-    # TODO: there is no error checking here to see that the user actually logged in successfully,
-    # and is not yet an active user.
-    if new_user is not None:
-        AUDIT_LOG.info(u"Login success on new account creation - {0}".format(new_user.username))
-
-    return new_user
-
-
-def skip_activation_email(user, do_external_auth, running_pipeline, third_party_provider):
-    """
-    Return `True` if activation email should be skipped.
-
-    Skip email if we are:
-        1. Doing load testing.
-        2. Random user generation for other forms of testing.
-        3. External auth bypassing activation.
-        4. Have the platform configured to not require e-mail activation.
-        5. Registering a new user using a trusted third party provider (with skip_email_verification=True)
-
-    Note that this feature is only tested as a flag set one way or
-    the other for *new* systems. we need to be careful about
-    changing settings on a running system to make sure no users are
-    left in an inconsistent state (or doing a migration if they are).
-
-    Arguments:
-        user (User): Django User object for the current user.
-        do_external_auth (bool): True if external authentication is in progress.
-        running_pipeline (dict): Dictionary containing user and pipeline data for third party authentication.
-        third_party_provider (ProviderConfig): An instance of third party provider configuration.
-
-    Returns:
-        (bool): `True` if account activation email should be skipped, `False` if account activation email should be
-            sent.
-    """
-    sso_pipeline_email = running_pipeline and running_pipeline['kwargs'].get('details', {}).get('email')
-
-    # Email is valid if the SAML assertion email matches the user account email or
-    # no email was provided in the SAML assertion. Some IdP's use a callback
-    # to retrieve additional user account information (including email) after the
-    # initial account creation.
-    valid_email = (
-        sso_pipeline_email == user.email or (
-            sso_pipeline_email is None and
-            third_party_provider and
-            getattr(third_party_provider, "identity_provider_type", None) == SAP_SUCCESSFACTORS_SAML_KEY
-        )
-    )
-
-    # log the cases where skip activation email flag is set, but email validity check fails
-    if third_party_provider and third_party_provider.skip_email_verification and not valid_email:
-        log.info(
-            '[skip_email_verification=True][user=%s][pipeline-email=%s][identity_provider=%s][provider_type=%s] '
-            'Account activation email sent as user\'s system email differs from SSO email.',
-            user.email,
-            sso_pipeline_email,
-            getattr(third_party_provider, "provider_id", None),
-            getattr(third_party_provider, "identity_provider_type", None)
-        )
-
-    return (
-        settings.FEATURES.get('SKIP_EMAIL_VALIDATION', None) or
-        settings.FEATURES.get('AUTOMATIC_AUTH_FOR_TESTING') or
-        (settings.FEATURES.get('BYPASS_ACTIVATION_EMAIL_FOR_EXTAUTH') and do_external_auth) or
-        (third_party_provider and third_party_provider.skip_email_verification and valid_email)
-    )
-
-
-def record_affiliate_registration_attribution(request, user):
-    """
-    Attribute this user's registration to the referring affiliate, if
-    applicable.
-    """
-    affiliate_id = request.COOKIES.get(settings.AFFILIATE_COOKIE_NAME)
-    if user and affiliate_id:
-        UserAttribute.set_user_attribute(user, REGISTRATION_AFFILIATE_ID, affiliate_id)
-
-
-def record_utm_registration_attribution(request, user):
-    """
-    Attribute this user's registration to the latest UTM referrer, if
-    applicable.
-    """
-    utm_cookie_name = RegistrationCookieConfiguration.current().utm_cookie_name
-    utm_cookie = request.COOKIES.get(utm_cookie_name)
-    if user and utm_cookie:
-        utm = json.loads(utm_cookie)
-        for utm_parameter_name in REGISTRATION_UTM_PARAMETERS:
-            utm_parameter = utm.get(utm_parameter_name)
-            if utm_parameter:
-                UserAttribute.set_user_attribute(
-                    user,
-                    REGISTRATION_UTM_PARAMETERS.get(utm_parameter_name),
-                    utm_parameter
-                )
-        created_at_unixtime = utm.get('created_at')
-        if created_at_unixtime:
-            # We divide by 1000 here because the javascript timestamp generated is in milliseconds not seconds.
-            # PYTHON: time.time()      => 1475590280.823698
-            # JS: new Date().getTime() => 1475590280823
-            created_at_datetime = datetime.datetime.fromtimestamp(int(created_at_unixtime) / float(1000), tz=UTC)
-            UserAttribute.set_user_attribute(
-                user,
-                REGISTRATION_UTM_CREATED_AT,
-                created_at_datetime
-            )
-
-
-def record_registration_attributions(request, user):
-    """
-    Attribute this user's registration based on referrer cookies.
-    """
-    record_affiliate_registration_attribution(request, user)
-    record_utm_registration_attribution(request, user)
-
-
-@csrf_exempt
-@transaction.non_atomic_requests
-def create_account(request, post_override=None):
-    """
-    JSON call to create new edX account.
-    Used by form in signup_modal.html, which is included into header.html
-    """
-    # Check if ALLOW_PUBLIC_ACCOUNT_CREATION flag turned off to restrict user account creation
-    if not configuration_helpers.get_value(
-            'ALLOW_PUBLIC_ACCOUNT_CREATION',
-            settings.FEATURES.get('ALLOW_PUBLIC_ACCOUNT_CREATION', True)
-    ):
-        return HttpResponseForbidden(_("Account creation not allowed."))
-
-    if waffle().is_enabled(PREVENT_AUTH_USER_WRITES):
-        return HttpResponseForbidden(SYSTEM_MAINTENANCE_MSG)
-
-    warnings.warn("Please use RegistrationView instead.", DeprecationWarning)
-
-    try:
-        user = create_account_with_params(request, post_override or request.POST)
-    except AccountValidationError as exc:
-        return JsonResponse({'success': False, 'value': text_type(exc), 'field': exc.field}, status=400)
-    except ValidationError as exc:
-        field, error_list = next(iteritems(exc.message_dict))
-        return JsonResponse(
-            {
-                "success": False,
-                "field": field,
-                "value": error_list[0],
-            },
-            status=400
-        )
-
-    redirect_url = None  # The AJAX method calling should know the default destination upon success
-
-    # Resume the third-party-auth pipeline if necessary.
-    if third_party_auth.is_enabled() and pipeline.running(request):
-        running_pipeline = pipeline.get(request)
-        redirect_url = pipeline.get_complete_url(running_pipeline['backend'])
-
-    response = JsonResponse({
-        'success': True,
-        'redirect_url': redirect_url,
-    })
-    set_logged_in_cookies(request, response, user)
-    return response
-
-
 @ensure_csrf_cookie
 def activate_account(request, key):
     """
@@ -1076,6 +641,83 @@ def activate_account_studio(request, key):
                 'already_active': already_active
             }
         )
+
+
+@require_http_methods(['POST'])
+def password_change_request_handler(request):
+    """Handle password change requests originating from the account page.
+
+    Uses the Account API to email the user a link to the password reset page.
+
+    Note:
+        The next step in the password reset process (confirmation) is currently handled
+        by student.views.password_reset_confirm_wrapper, a custom wrapper around Django's
+        password reset confirmation view.
+
+    Args:
+        request (HttpRequest)
+
+    Returns:
+        HttpResponse: 200 if the email was sent successfully
+        HttpResponse: 400 if there is no 'email' POST parameter
+        HttpResponse: 403 if the client has been rate limited
+        HttpResponse: 405 if using an unsupported HTTP method
+
+    Example usage:
+
+        POST /account/password
+
+    """
+
+    limiter = BadRequestRateLimiter()
+    if limiter.is_rate_limit_exceeded(request):
+        AUDIT_LOG.warning("Password reset rate limit exceeded")
+        return HttpResponseForbidden()
+
+    user = request.user
+    # Prefer logged-in user's email
+    email = user.email if user.is_authenticated else request.POST.get('email')
+
+    if email:
+        try:
+            from openedx.core.djangoapps.user_api.accounts.api import request_password_change
+            request_password_change(email, request.is_secure())
+            user = user if user.is_authenticated else User.objects.get(email=email)
+            destroy_oauth_tokens(user)
+        except UserNotFound:
+            AUDIT_LOG.info("Invalid password reset attempt")
+            # Increment the rate limit counter
+            limiter.tick_bad_request_counter(request)
+
+            # If enabled, send an email saying that a password reset was attempted, but that there is
+            # no user associated with the email
+            if configuration_helpers.get_value('ENABLE_PASSWORD_RESET_FAILURE_EMAIL',
+                                               settings.FEATURES['ENABLE_PASSWORD_RESET_FAILURE_EMAIL']):
+
+                site = get_current_site()
+                message_context = get_base_template_context(site)
+
+                message_context.update({
+                    'failed': True,
+                    'request': request,  # Used by google_analytics_tracking_pixel
+                    'email_address': email,
+                })
+
+                msg = PasswordReset().personalize(
+                    recipient=Recipient(username='', email_address=email),
+                    language=settings.LANGUAGE_CODE,
+                    user_context=message_context,
+                )
+
+                ace.send(msg)
+        except UserAPIInternalError as err:
+            log.exception('Error occured during password change for user {email}: {error}'
+                          .format(email=email, error=err))
+            return HttpResponse(_("Some error occured during password change. Please try again"), status=500)
+
+        return HttpResponse(status=200)
+    else:
+        return HttpResponseBadRequest(_("No email address provided."))
 
 
 @csrf_exempt
