+++ common/djangoapps/third_party_auth/tests/specs/base.py
@@ -20,11 +20,12 @@ from social_django import utils as social_utils
 from social_django import views as social_views
 
 from lms.djangoapps.commerce.tests import TEST_API_URL
+from openedx.core.djangoapps.user_authn.views.deprecated import signin_user, create_account, register_user
+from openedx.core.djangoapps.user_authn.views.login import login_user
 from openedx.core.djangoapps.site_configuration.tests.factories import SiteFactory
+from openedx.core.djangoapps.user_api.accounts.settings_views import account_settings_context
 from student import models as student_models
-from student import views as student_views
 from student.tests.factories import UserFactory
-from student_account.views import account_settings_context
 
 from third_party_auth import middleware, pipeline
 from third_party_auth.tests import testutil
@@ -538,8 +539,8 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
         actions.do_complete(request.backend, social_views._do_login,  # pylint: disable=protected-access
                             request=request)
 
-        student_views.signin_user(strategy.request)
-        student_views.login_user(strategy.request)
+        signin_user(strategy.request)
+        login_user(strategy.request)
         actions.do_complete(request.backend, social_views._do_login,  # pylint: disable=protected-access
                             request=request)
 
@@ -595,8 +596,8 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
                             request=request)
 
         with self._patch_edxmako_current_request(strategy.request):
-            student_views.signin_user(strategy.request)
-            student_views.login_user(strategy.request)
+            signin_user(strategy.request)
+            login_user(strategy.request)
             actions.do_complete(request.backend, social_views._do_login, user=user,  # pylint: disable=protected-access
                                 request=request)
 
@@ -662,8 +663,8 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
                             request=request)
 
         with self._patch_edxmako_current_request(strategy.request):
-            student_views.signin_user(strategy.request)
-            student_views.login_user(strategy.request)
+            signin_user(strategy.request)
+            login_user(strategy.request)
             actions.do_complete(request.backend, social_views._do_login,  # pylint: disable=protected-access
                                 user=user, request=request)
 
@@ -707,12 +708,12 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
         # At this point we know the pipeline has resumed correctly. Next we
         # fire off the view that displays the login form and posts it via JS.
         with self._patch_edxmako_current_request(strategy.request):
-            self.assert_login_response_in_pipeline_looks_correct(student_views.signin_user(strategy.request))
+            self.assert_login_response_in_pipeline_looks_correct(signin_user(strategy.request))
 
         # Next, we invoke the view that handles the POST, and expect it
         # redirects to /auth/complete. In the browser ajax handlers will
         # redirect the user to the dashboard; we invoke it manually here.
-        self.assert_json_success_response_looks_correct(student_views.login_user(strategy.request))
+        self.assert_json_success_response_looks_correct(login_user(strategy.request))
 
         # We should be redirected back to the complete page, setting
         # the "logged in" cookie for the marketing site.
@@ -739,7 +740,7 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
         user.save()
 
         with self._patch_edxmako_current_request(strategy.request):
-            self.assert_json_failure_response_is_inactive_account(student_views.login_user(strategy.request))
+            self.assert_json_failure_response_is_inactive_account(login_user(strategy.request))
 
     def test_signin_fails_if_no_account_associated(self):
         _, strategy = self.get_request_and_strategy(
@@ -748,7 +749,7 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
         self.create_user_models_for_existing_account(
             strategy, 'user@example.com', 'password', self.get_username(), skip_social_auth=True)
 
-        self.assert_json_failure_response_is_missing_social_auth(student_views.login_user(strategy.request))
+        self.assert_json_failure_response_is_missing_social_auth(login_user(strategy.request))
 
     def test_first_party_auth_trumps_third_party_auth_but_is_invalid_when_only_email_in_request(self):
         self.assert_first_party_auth_trumps_third_party_auth(email='user@example.com')
@@ -789,7 +790,7 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
         # fire off the view that displays the registration form.
         with self._patch_edxmako_current_request(request):
             self.assert_register_response_in_pipeline_looks_correct(
-                student_views.register_user(strategy.request),
+                register_user(strategy.request),
                 pipeline.get(request)['kwargs'],
                 ['name', 'username', 'email']
             )
@@ -811,7 +812,7 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
         # ...but when we invoke create_account the existing edX view will make
         # it, but not social auths. The pipeline creates those later.
         with self._patch_edxmako_current_request(strategy.request):
-            self.assert_json_success_response_looks_correct(student_views.create_account(strategy.request))
+            self.assert_json_success_response_looks_correct(create_account(strategy.request))
         # We've overridden the user's password, so authenticate() with the old
         # value won't work:
         created_user = self.get_user_by_email(strategy, email)
@@ -864,7 +865,7 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
 
         with self._patch_edxmako_current_request(request):
             self.assert_register_response_in_pipeline_looks_correct(
-                student_views.register_user(strategy.request),
+                register_user(strategy.request),
                 pipeline.get(request)['kwargs'],
                 ['name', 'username', 'email']
             )
@@ -872,8 +873,8 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
         with self._patch_edxmako_current_request(strategy.request):
             strategy.request.POST = self.get_registration_post_vars()
             # Create twice: once successfully, and once causing a collision.
-            student_views.create_account(strategy.request)
-        self.assert_json_failure_response_is_username_collision(student_views.create_account(strategy.request))
+            create_account(strategy.request)
+        self.assert_json_failure_response_is_username_collision(create_account(strategy.request))
 
     def test_pipeline_raises_auth_entry_error_if_auth_entry_invalid(self):
         auth_entry = 'invalid'
@@ -914,7 +915,7 @@ class IntegrationTest(testutil.TestCase, test.TestCase, HelperMixin):
             strategy.request.POST['password'] = 'bad_' + password if success is False else password
 
         self.assert_pipeline_running(strategy.request)
-        payload = json.loads(student_views.login_user(strategy.request).content)
+        payload = json.loads(login_user(strategy.request).content)
 
         if success is None:
             # Request malformed -- just one of email/password given.
