+++ common/djangoapps/third_party_auth/tests/specs/test_testshib.py
@@ -18,9 +18,10 @@ from social_django.models import UserSocialAuth
 from testfixtures import LogCapture
 
 from enterprise.models import EnterpriseCustomerIdentityProvider, EnterpriseCustomerUser
+from openedx.core.djangoapps.user_authn.views.deprecated import signin_user
+from openedx.core.djangoapps.user_authn.views.login import login_user
+from openedx.core.djangoapps.user_api.accounts.settings_views import account_settings_context
 from openedx.features.enterprise_support.tests.factories import EnterpriseCustomerFactory
-from student import views as student_views
-from student_account.views import account_settings_context
 from third_party_auth import pipeline
 from third_party_auth.saml import SapSuccessFactorsIdentityProvider, log as saml_log
 from third_party_auth.tasks import fetch_saml_metadata
@@ -183,8 +184,8 @@ class TestShibIntegrationTest(SamlIntegrationTestUtilities, IntegrationTestMixin
                             request=request)
 
         with self._patch_edxmako_current_request(strategy.request):
-            student_views.signin_user(strategy.request)
-            student_views.login_user(strategy.request)
+            signin_user(strategy.request)
+            login_user(strategy.request)
             actions.do_complete(request.backend, social_views._do_login, user=user,  # pylint: disable=protected-access
                                 request=request)
