+++ lms/djangoapps/email_marketing/signals.py
@@ -17,10 +17,10 @@ import third_party_auth
 from course_modes.models import CourseMode
 from email_marketing.models import EmailMarketingConfiguration
 from lms.djangoapps.email_marketing.tasks import get_email_cookies_via_sailthru, update_user, update_user_email
+from openedx.core.djangoapps.user_authn.cookies import CREATE_LOGON_COOKIE
 from openedx.core.djangoapps.lang_pref import LANGUAGE_KEY
 from openedx.core.djangoapps.user_api.accounts.signals import USER_RETIRE_THIRD_PARTY_MAILINGS
 from openedx.core.djangoapps.waffle_utils import WaffleSwitchNamespace
-from student.cookies import CREATE_LOGON_COOKIE
 from student.signals import SAILTHRU_AUDIT_PURCHASE
 from student.views import REGISTER_USER
 from util.model_utils import USER_FIELD_CHANGED
