+++ openedx/core/djangoapps/user_authn/cookies.py
