+++ openedx/core/djangoapps/user_authn/tests/test_cookies.py
@@ -6,8 +6,8 @@ from django.conf import settings
 from django.urls import reverse
 from django.test import RequestFactory
 
+from openedx.core.djangoapps.user_authn.cookies import get_user_info_cookie_data
 from openedx.core.djangoapps.user_api.accounts.utils import retrieve_last_sitewide_block_completed
-from student.cookies import get_user_info_cookie_data
 from student.models import CourseEnrollment
 from student.tests.factories import UserFactory
 from xmodule.modulestore.tests.django_utils import SharedModuleStoreTestCase
