+++ openedx/core/djangoapps/user_authn/views/tests/test_views.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
-""" Tests for student account views. """
+""" Tests for user authn views. """
 
+from http.cookies import SimpleCookie
 import logging
 import re
 from unittest import skipUnless
@@ -17,14 +18,11 @@ from django.contrib.sessions.middleware import SessionMiddleware
 from django.core import mail
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.urls import reverse
-from django.http import HttpRequest
 from django.test import TestCase
 from django.test.client import RequestFactory
 from django.test.utils import override_settings
 from django.utils.translation import ugettext as _
 from edx_oauth2_provider.tests.factories import AccessTokenFactory, ClientFactory, RefreshTokenFactory
-from edx_rest_api_client import exceptions
-from http.cookies import SimpleCookie
 from oauth2_provider.models import AccessToken as dot_access_token
 from oauth2_provider.models import RefreshToken as dot_refresh_token
 from provider.oauth2.models import AccessToken as dop_access_token
@@ -32,25 +30,18 @@ from provider.oauth2.models import RefreshToken as dop_refresh_token
 from testfixtures import LogCapture
 
 from course_modes.models import CourseMode
-from lms.djangoapps.commerce.models import CommerceConfiguration
-from lms.djangoapps.commerce.tests import factories
-from lms.djangoapps.commerce.tests.mocks import mock_get_orders
-from lms.djangoapps.student_account.views import login_and_registration_form
+from openedx.core.djangoapps.user_authn.views.login_form import login_and_registration_form
 from openedx.core.djangoapps.oauth_dispatch.tests import factories as dot_factories
-from openedx.core.djangoapps.programs.tests.mixins import ProgramsApiConfigMixin
-from openedx.core.djangoapps.site_configuration.tests.factories import SiteFactory
 from openedx.core.djangoapps.site_configuration.tests.mixins import SiteMixin
 from openedx.core.djangoapps.theming.tests.test_util import with_comprehensive_theme_context
 from openedx.core.djangoapps.user_api.accounts.api import activate_account, create_account
+from openedx.core.djangoapps.user_api.errors import UserAPIInternalError
 from openedx.core.djangolib.js_utils import dump_js_escaped_json
 from openedx.core.djangolib.markup import HTML, Text
-from openedx.core.djangolib.testing.utils import CacheIsolationTestCase
-from student.tests.factories import UserFactory
-from student_account.views import account_settings_context, get_user_orders
+from openedx.core.djangolib.testing.utils import CacheIsolationTestCase, skip_unless_lms
 from third_party_auth.tests.testutil import ThirdPartyAuthTestMixin, simulate_running_pipeline
 from util.testing import UrlResetMixin
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
-from openedx.core.djangoapps.user_api.errors import UserAPIInternalError
 
 LOGGER_NAME = 'audit'
 User = get_user_model()  # pylint:disable=invalid-name
@@ -59,9 +50,10 @@ FEATURES_WITH_FAILED_PASSWORD_RESET_EMAIL = settings.FEATURES.copy()
 FEATURES_WITH_FAILED_PASSWORD_RESET_EMAIL['ENABLE_PASSWORD_RESET_FAILURE_EMAIL'] = True
 
 
+@skip_unless_lms
 @ddt.ddt
-class StudentAccountUpdateTest(CacheIsolationTestCase, UrlResetMixin):
-    """ Tests for the student account views that update the user's account information. """
+class UserAccountUpdateTest(CacheIsolationTestCase, UrlResetMixin):
+    """ Tests for views that update the user's account information. """
 
     USERNAME = u"heisenberg"
     ALTERNATE_USERNAME = u"walt"
@@ -78,7 +70,7 @@ class StudentAccountUpdateTest(CacheIsolationTestCase, UrlResetMixin):
     ENABLED_CACHES = ['default']
 
     def setUp(self):
-        super(StudentAccountUpdateTest, self).setUp()
+        super(UserAccountUpdateTest, self).setUp()
 
         # Create/activate a new account
         activation_key = create_account(self.USERNAME, self.OLD_PASSWORD, self.OLD_EMAIL)
@@ -293,8 +285,9 @@ class StudentAccountUpdateTest(CacheIsolationTestCase, UrlResetMixin):
         self.assertFalse(dop_refresh_token.objects.filter(user=user).exists())
 
 
+@skip_unless_lms
 @ddt.ddt
-class StudentAccountLoginAndRegistrationTest(ThirdPartyAuthTestMixin, UrlResetMixin, ModuleStoreTestCase):
+class LoginAndRegistrationTest(ThirdPartyAuthTestMixin, UrlResetMixin, ModuleStoreTestCase):
     """ Tests for the student account views that update the user's account information. """
     shard = 7
     USERNAME = "bob"
@@ -304,8 +297,8 @@ class StudentAccountLoginAndRegistrationTest(ThirdPartyAuthTestMixin, UrlResetMi
     URLCONF_MODULES = ['openedx.core.djangoapps.embargo']
 
     @mock.patch.dict(settings.FEATURES, {'EMBARGO': True})
-    def setUp(self):
-        super(StudentAccountLoginAndRegistrationTest, self).setUp()
+    def setUp(self):  # pylint: disable=arguments-differ
+        super(LoginAndRegistrationTest, self).setUp()
 
         # Several third party auth providers are created for these tests:
         self.google_provider = self.configure_google_provider(enabled=True, visible=True)
@@ -387,7 +380,7 @@ class StudentAccountLoginAndRegistrationTest(ThirdPartyAuthTestMixin, UrlResetMi
         response = self.client.get(reverse(url_name))
         self._assert_third_party_auth_data(response, None, None, [], None)
 
-    @mock.patch('student_account.views.enterprise_customer_for_request')
+    @mock.patch('openedx.core.djangoapps.user_authn.views.login_form.enterprise_customer_for_request')
     @mock.patch('openedx.core.djangoapps.user_api.api.enterprise_customer_for_request')
     @ddt.data(
         ("signin_user", None, None, None, False),
@@ -448,7 +441,7 @@ class StudentAccountLoginAndRegistrationTest(ThirdPartyAuthTestMixin, UrlResetMi
 
         # Simulate a running pipeline
         if current_backend is not None:
-            pipeline_target = "student_account.views.third_party_auth.pipeline"
+            pipeline_target = "openedx.core.djangoapps.user_authn.views.login_form.third_party_auth.pipeline"
             with simulate_running_pipeline(pipeline_target, current_backend, email=email):
                 response = self.client.get(reverse(url_name), params, HTTP_ACCEPT="text/html")
 
@@ -509,7 +502,7 @@ class StudentAccountLoginAndRegistrationTest(ThirdPartyAuthTestMixin, UrlResetMi
         self.configure_saml_provider(**kwargs)
 
     @mock.patch('django.conf.settings.MESSAGE_STORAGE', 'django.contrib.messages.storage.cookie.CookieStorage')
-    @mock.patch('lms.djangoapps.student_account.views.enterprise_customer_for_request')
+    @mock.patch('openedx.core.djangoapps.user_authn.views.login_form.enterprise_customer_for_request')
     @ddt.data(
         (
             'signin_user',
@@ -552,7 +545,7 @@ class StudentAccountLoginAndRegistrationTest(ThirdPartyAuthTestMixin, UrlResetMi
                 'idp_name': dummy_idp
             }
         }
-        pipeline_target = 'student_account.views.third_party_auth.pipeline'
+        pipeline_target = 'openedx.core.djangoapps.user_authn.views.login_form.third_party_auth.pipeline'
         with simulate_running_pipeline(pipeline_target, current_backend, **pipeline_response):
             with mock.patch('edxmako.request_context.get_current_request', return_value=request):
                 response = login_and_registration_form(request)
@@ -653,7 +646,7 @@ class StudentAccountLoginAndRegistrationTest(ThirdPartyAuthTestMixin, UrlResetMi
             target_status_code=302
         )
 
-    @mock.patch('student_account.views.enterprise_customer_for_request')
+    @mock.patch('openedx.core.djangoapps.user_authn.views.login_form.enterprise_customer_for_request')
     @ddt.data(
         ('signin_user', False, None, None),
         ('register_user', False, None, None),
@@ -850,214 +843,7 @@ class StudentAccountLoginAndRegistrationTest(ThirdPartyAuthTestMixin, UrlResetMi
         self.assertEqual(response['Content-Language'], 'es-es')
 
 
-class AccountSettingsViewTest(ThirdPartyAuthTestMixin, TestCase, ProgramsApiConfigMixin):
-    """ Tests for the account settings view. """
-
-    USERNAME = 'student'
-    PASSWORD = 'password'
-    FIELDS = [
-        'country',
-        'gender',
-        'language',
-        'level_of_education',
-        'password',
-        'year_of_birth',
-        'preferred_language',
-        'time_zone',
-    ]
-
-    @mock.patch("django.conf.settings.MESSAGE_STORAGE", 'django.contrib.messages.storage.cookie.CookieStorage')
-    def setUp(self):
-        super(AccountSettingsViewTest, self).setUp()
-        self.user = UserFactory.create(username=self.USERNAME, password=self.PASSWORD)
-        CommerceConfiguration.objects.create(cache_ttl=10, enabled=True)
-        self.client.login(username=self.USERNAME, password=self.PASSWORD)
-
-        self.request = HttpRequest()
-        self.request.user = self.user
-
-        # For these tests, two third party auth providers are enabled by default:
-        self.configure_google_provider(enabled=True, visible=True)
-        self.configure_facebook_provider(enabled=True, visible=True)
-
-        # Python-social saves auth failure notifcations in Django messages.
-        # See pipeline.get_duplicate_provider() for details.
-        self.request.COOKIES = {}
-        MessageMiddleware().process_request(self.request)
-        messages.error(self.request, 'Facebook is already in use.', extra_tags='Auth facebook')
-
-    @mock.patch('openedx.features.enterprise_support.api.get_enterprise_customer_for_learner')
-    def test_context(self, mock_get_enterprise_customer_for_learner):
-        self.request.site = SiteFactory.create()
-        mock_get_enterprise_customer_for_learner.return_value = {}
-        context = account_settings_context(self.request)
-
-        user_accounts_api_url = reverse("accounts_api", kwargs={'username': self.user.username})
-        self.assertEqual(context['user_accounts_api_url'], user_accounts_api_url)
-
-        user_preferences_api_url = reverse('preferences_api', kwargs={'username': self.user.username})
-        self.assertEqual(context['user_preferences_api_url'], user_preferences_api_url)
-
-        for attribute in self.FIELDS:
-            self.assertIn(attribute, context['fields'])
-
-        self.assertEqual(
-            context['user_accounts_api_url'], reverse("accounts_api", kwargs={'username': self.user.username})
-        )
-        self.assertEqual(
-            context['user_preferences_api_url'], reverse('preferences_api', kwargs={'username': self.user.username})
-        )
-
-        self.assertEqual(context['duplicate_provider'], 'facebook')
-        self.assertEqual(context['auth']['providers'][0]['name'], 'Facebook')
-        self.assertEqual(context['auth']['providers'][1]['name'], 'Google')
-
-        self.assertEqual(context['sync_learner_profile_data'], False)
-        self.assertEqual(context['edx_support_url'], settings.SUPPORT_SITE_LINK)
-        self.assertEqual(context['enterprise_name'], None)
-        self.assertEqual(
-            context['enterprise_readonly_account_fields'], {'fields': settings.ENTERPRISE_READONLY_ACCOUNT_FIELDS}
-        )
-
-    @mock.patch('student_account.views.get_enterprise_customer_for_learner')
-    @mock.patch('openedx.features.enterprise_support.utils.third_party_auth.provider.Registry.get')
-    def test_context_for_enterprise_learner(
-            self, mock_get_auth_provider, mock_get_enterprise_customer_for_learner
-    ):
-        dummy_enterprise_customer = {
-            'uuid': 'real-ent-uuid',
-            'name': 'Dummy Enterprise',
-            'identity_provider': 'saml-ubc'
-        }
-        mock_get_enterprise_customer_for_learner.return_value = dummy_enterprise_customer
-        self.request.site = SiteFactory.create()
-        mock_get_auth_provider.return_value.sync_learner_profile_data = True
-        context = account_settings_context(self.request)
-
-        user_accounts_api_url = reverse("accounts_api", kwargs={'username': self.user.username})
-        self.assertEqual(context['user_accounts_api_url'], user_accounts_api_url)
-
-        user_preferences_api_url = reverse('preferences_api', kwargs={'username': self.user.username})
-        self.assertEqual(context['user_preferences_api_url'], user_preferences_api_url)
-
-        for attribute in self.FIELDS:
-            self.assertIn(attribute, context['fields'])
-
-        self.assertEqual(
-            context['user_accounts_api_url'], reverse("accounts_api", kwargs={'username': self.user.username})
-        )
-        self.assertEqual(
-            context['user_preferences_api_url'], reverse('preferences_api', kwargs={'username': self.user.username})
-        )
-
-        self.assertEqual(context['duplicate_provider'], 'facebook')
-        self.assertEqual(context['auth']['providers'][0]['name'], 'Facebook')
-        self.assertEqual(context['auth']['providers'][1]['name'], 'Google')
-
-        self.assertEqual(
-            context['sync_learner_profile_data'], mock_get_auth_provider.return_value.sync_learner_profile_data
-        )
-        self.assertEqual(context['edx_support_url'], settings.SUPPORT_SITE_LINK)
-        self.assertEqual(context['enterprise_name'], dummy_enterprise_customer['name'])
-        self.assertEqual(
-            context['enterprise_readonly_account_fields'], {'fields': settings.ENTERPRISE_READONLY_ACCOUNT_FIELDS}
-        )
-
-    def test_view(self):
-        """
-        Test that all fields are  visible
-        """
-        view_path = reverse('account_settings')
-        response = self.client.get(path=view_path)
-
-        for attribute in self.FIELDS:
-            self.assertIn(attribute, response.content)
-
-    def test_header_with_programs_listing_enabled(self):
-        """
-        Verify that tabs header will be shown while program listing is enabled.
-        """
-        self.create_programs_config()
-        view_path = reverse('account_settings')
-        response = self.client.get(path=view_path)
-
-        self.assertContains(response, 'global-header')
-
-    def test_header_with_programs_listing_disabled(self):
-        """
-        Verify that nav header will be shown while program listing is disabled.
-        """
-        self.create_programs_config(enabled=False)
-        view_path = reverse('account_settings')
-        response = self.client.get(path=view_path)
-
-        self.assertContains(response, 'global-header')
-
-    def test_commerce_order_detail(self):
-        """
-        Verify that get_user_orders returns the correct order data.
-        """
-        with mock_get_orders():
-            order_detail = get_user_orders(self.user)
-
-        for i, order in enumerate(mock_get_orders.default_response['results']):
-            expected = {
-                'number': order['number'],
-                'price': order['total_excl_tax'],
-                'order_date': 'Jan 01, 2016',
-                'receipt_url': '/checkout/receipt/?order_number=' + order['number'],
-                'lines': order['lines'],
-            }
-            self.assertEqual(order_detail[i], expected)
-
-    def test_commerce_order_detail_exception(self):
-        with mock_get_orders(exception=exceptions.HttpNotFoundError):
-            order_detail = get_user_orders(self.user)
-
-        self.assertEqual(order_detail, [])
-
-    def test_incomplete_order_detail(self):
-        response = {
-            'results': [
-                factories.OrderFactory(
-                    status='Incomplete',
-                    lines=[
-                        factories.OrderLineFactory(
-                            product=factories.ProductFactory(attribute_values=[factories.ProductAttributeFactory()])
-                        )
-                    ]
-                )
-            ]
-        }
-        with mock_get_orders(response=response):
-            order_detail = get_user_orders(self.user)
-
-        self.assertEqual(order_detail, [])
-
-    def test_order_history_with_no_product(self):
-        response = {
-            'results': [
-                factories.OrderFactory(
-                    lines=[
-                        factories.OrderLineFactory(
-                            product=None
-                        ),
-                        factories.OrderLineFactory(
-                            product=factories.ProductFactory(attribute_values=[factories.ProductAttributeFactory(
-                                name='certificate_type',
-                                value='verified'
-                            )])
-                        )
-                    ]
-                )
-            ]
-        }
-        with mock_get_orders(response=response):
-            order_detail = get_user_orders(self.user)
-
-        self.assertEqual(len(order_detail), 1)
-
-
+@skip_unless_lms
 @override_settings(SITE_NAME=settings.MICROSITE_LOGISTRATION_HOSTNAME)
 class MicrositeLogistrationTests(TestCase):
     """
@@ -1115,6 +901,7 @@ class MicrositeLogistrationTests(TestCase):
         self.assertNotIn('<div id="login-and-registration-container"', resp.content)
 
 
+@skip_unless_lms
 class AccountCreationTestCaseWithSiteOverrides(SiteMixin, TestCase):
     """
     Test cases for Feature flag ALLOW_PUBLIC_ACCOUNT_CREATION which when
