+++ openedx/features/enterprise_support/utils.py
@@ -8,8 +8,8 @@ from django.utils.translation import ugettext as _
 
 import third_party_auth
 from third_party_auth import pipeline
-from student.cookies import set_experiments_is_enterprise_cookie
 
+from openedx.core.djangoapps.user_authn.cookies import set_experiments_is_enterprise_cookie
 from openedx.core.djangoapps.site_configuration import helpers as configuration_helpers
 from openedx.core.djangolib.markup import HTML, Text
